AD-FI (eclOut2mrst): Fix Type Error
This function expects to treat 'actNum' as LOGICAL, so don't
convert to INT32 except when forming 'grdecl' structure.

Prompted by issue in MRST User Forum.

+++ ad-fi/ecl/eclOut2mrst.m
@@ -16,13 +16,11 @@ else
     zcorn  = grid.ZCORN.values;
 end
 
-actNum   = int32(and(anGrid, init.PORV.values > 0));
-
-
+actNum = and(anGrid, init.PORV.values > 0);
 grdecl = struct('cartDims', cartDims, ...
                 'COORD'   , convertFrom(coord, u.length), ...
                 'ZCORN'   , convertFrom(zcorn, u.length), ...
-                'ACTNUM'  , actNum );
+                'ACTNUM'  , int32(actNum));
 try 
     mrstModule add libgeometry mex opm_gridprocessing
     G = mprocessGRDECL(grdecl, 'SplitDisconnected', false);
